Modified dump:
Dump is no longer the function for eigenvector or Pearson's, as these are not really dump functions but rather calculation functions.

Dump no longer takes the unnecessary second chromosome for norm since it is ignored.

Pearson and Eigenvector are now their own calls and can be called on any resolution.

Removed and cleaned up redundant code.