Changed Feature2DListHandler So That it can only contain one Feature2D List
Previously Feature2DListHandler allowed for multiple Feature2DLists. This was used in a previous implementation of annotation layers that had Feature2DListHandler controlling the layers. This implementation is now handled by AnnotationLayerHandler and as such this is no longer needed.

This issue caused for multiple lists to be created when resizing and creating new annotations. It also made deleting some annotations impossible due to the list that contained them not being the list that was intended.

As a result of this change: Resolved #522  #524  #525  #528

 In addition to the changes above the function getActiveLayer() was changed to getActiveLayerHandler() for better readability